+++ src/distfuns.cpp
@@ -2,11 +2,11 @@
 #include <cmath>
 #include "misc.h"
 
-
 // via the depends attribute we tell Rcpp to create hooks for
 // RcppArmadillo so that the build process will know what to do
 //
 // [[Rcpp::depends(RcppArmadillo)]]
+
 
 //' Get the distances for computing the partition function given
 //' the cardinalities.
@@ -14,7 +14,7 @@
 //' @param n The number of items.
 //' @param cardinalities Number of times each distance appears.
 //' @param metric A string. Available options are \code{"footrule"},
-//' \code{"kendall"}, and \code{"spearman"}. Defaults to \code{"footrule"}.
+//' and \code{"spearman"}. Defaults to \code{"footrule"}.
 //' @return A scalar, the logarithm of the partition function.
 //' @keywords internal
 //'
@@ -52,7 +52,8 @@ arma::vec get_summation_distances(int n, arma::vec cardinalities,
 //' @param r1 A vector of ranks.
 //' @param r2 A vector of ranks.
 //' @param metric A string. Available options are \code{"footrule"},
-//' \code{"kendall"}, \code{"cayley"}, \code{"hamming"} and \code{"spearman"}. Defaults to \code{"footrule"}.
+//' \code{"kendall"}, \code{"cayley"}, \code{"hamming"}, and \code{"spearman"}.
+//' Defaults to \code{"footrule"}.
 //' @return A scalar.
 //' @details Note that the Spearman distance is the squared L2 norm, whereas
 //' the footrule distance is the L1 norm.
@@ -140,3 +141,7 @@ void update_distance_matrix(arma::mat& dist_mat, const arma::mat& rankings, cons
   }
 
 }
+
+
+
+