+++ tests/testthat/test_partition_function.R
@@ -22,6 +22,10 @@ check_log_zn <- function(n, alpha, metric){
     log(sum(exp(- alpha / n * apply(perm, 1,
                                     get_rank_distance,
                                     r2 = 1:n, metric = "hamming"))))
+  } else if(metric == "ulam") {
+    log(sum(unlist(lapply(seq(0, n - 1, by = 1), function(x) {
+      PerMallows::count.perms(perm.length = n, dist.value = x, dist.name = "ulam") * exp(-alpha / n * x)
+    }))))
   } else {
     stop("Unknown metric.")
   }
@@ -89,3 +93,4 @@ test_that("Hamming partition function is correct", {
       )
     }
   }})
+